# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tropski bar",
        "description": "Swagger - tropski bar",
        "contact": {
            "email": "dtahirovic2@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Find category by ID",
                "description": "Returns a single category",
                "operationId": "categories.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid category ID supplied"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update an existing category",
                "description": "",
                "operationId": "categories.update",
                "requestBody": {
                    "description": "Category object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Deletes a category",
                "description": "",
                "operationId": "categories.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
        "/api/category-trans/{id}": {
            "get": {
                "tags": [
                    "CategoryTran"
                ],
                "summary": "Find category translation by ID",
                "description": "Returns a single category translation",
                "operationId": "category-trans.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category translation to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid category translation ID supplied"
                    },
                    "404": {
                        "description": "Category translation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "CategoryTran"
                ],
                "summary": "Update an existing category translation",
                "description": "",
                "operationId": "category-trans.update",
                "requestBody": {
                    "description": "Category translation object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryTran"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Category translation not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CategoryTran"
                ],
                "summary": "Deletes a category translation",
                "description": "",
                "operationId": "category-trans.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category translation id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Category translation not found"
                    }
                }
            }
        },
...
        "/api/extras/{id}": {
            "get": {
                "tags": [
                    "Extra"
                ],
                "summary": "Find extra by ID",
                "description": "Returns a single extra",
                "operationId": "extras.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of extra to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid extra ID supplied"
                    },
                    "404": {
                        "description": "Extra not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Extra"
                ],
                "summary": "Update an existing extra",
                "description": "",
                "operationId": "extras.update",
                "requestBody": {
                    "description": "Extra object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Extra"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Extra"
                ],
                "summary": "Deletes an extra",
                "description": "",
                "operationId": "extras.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Extra id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra not found"
                    }
                }
            }
        },
...
        "/api/extra-groups/{id}": {
            "get": {
                "tags": [
                    "ExtraGroup"
                ],
                "summary": "Find extra group by ID",
                "description": "Returns a single extra group",
                "operationId": "extra-groups.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of extra group to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid extra group ID supplied"
                    },
                    "404": {
                        "description": "Extra group not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ExtraGroup"
                ],
                "summary": "Update an existing extra group",
                "description": "",
                "operationId": "extra-groups.update",
                "requestBody": {
                    "description": "Extra group object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExtraGroup"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra group not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ExtraGroup"
                ],
                "summary": "Deletes an extra group",
                "description": "",
                "operationId": "extra-groups.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Extra group id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra group not found"
                    }
                }
            }
        },
...
        "/api/extra-group-extra-pivots/{id}": {
            "get": {
                "tags": [
                    "ExtraGroupExtraPivot"
                ],
                "summary": "Find extra group pivot by ID",
                "description": "Returns a single extra group pivot",
                "operationId": "extra-group-extra-pivots.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of extra group pivot to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid extra group pivot ID supplied"
                    },
                    "404": {
                        "description": "Extra group pivot not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ExtraGroupExtraPivot"
                ],
                "summary": "Update an existing extra group pivot",
                "description": "",
                "operationId": "extra-group-extra-pivots.update",
                "requestBody": {
                    "description": "Extra group pivot object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExtraGroupExtraPivot"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra group pivot not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ExtraGroupExtraPivot"
                ],
                "summary": "Deletes an extra group pivot",
                "description": "",
                "operationId": "extra-group-extra-pivots.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Extra group pivot id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra group pivot not found"
                    }
                }
            }
        },
...
        "/api/extra-trans/{id}": {
            "get": {
                "tags": [
                    "ExtraTran"
                ],
                "summary": "Find extra translation by ID",
                "description": "Returns a single extra translation",
                "operationId": "extra-trans.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of extra translation to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid extra translation ID supplied"
                    },
                    "404": {
                        "description": "Extra translation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "ExtraTran"
                ],
                "summary": "Update an existing extra translation",
                "description": "",
                "operationId": "extra-trans.update",
                "requestBody": {
                    "description": "Extra translation object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExtraTran"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra translation not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ExtraTran"
                ],
                "summary": "Deletes an extra translation",
                "description": "",
                "operationId": "extra-trans.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Extra translation id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Extra translation not found"
                    }
                }
            }
        },
...
        "/api/hotel-users/{id}": {
            "get": {
                "tags": [
                    "HotelUser"
                ],
                "summary": "Find hotel user by ID",
                "description": "Returns a single hotel user",
                "operationId": "hotel-users.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of hotel user to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid hotel user ID supplied"
                    },
                    "404": {
                        "description": "Hotel user not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "HotelUser"
                ],
                "summary": "Update an existing hotel user",
                "description": "",
                "operationId": "hotel-users.update",
                "requestBody": {
                    "description": "Hotel user object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HotelUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Hotel user not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "HotelUser"
                ],
                "summary": "Deletes a hotel user",
                "description": "",
                "operationId": "hotel-users.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hotel user id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Hotel user not found"
                    }
                }
            }
        },
...
        "/api/item-types": {
            "get": {
                "tags": [
                    "ItemType"
                ],
                "summary": "Finds all item types",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "item-types.index",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "active",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "ItemType"
                ],
                "summary": "Create a new item type",
                "operationId": "item-types.store",
                "requestBody": {
                    "description": "Item type data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItemType"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item type created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "required": [
                    "name_en",
                    "name_me",
                    "main_cat_id"
                ],
                "properties": {
                    "name_en": {
                        "description": "The English name of the category",
                        "type": "string"
                    },
                    "name_me": {
                        "description": "The Montenegrin name of the category",
                        "type": "string"
                    },
                    "main_cat_id": {
                        "description": "The ID of the main category",
                        "type": "integer"
                    },
                    "image": {
                        "description": "The image of the category",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
            "CategoryTran": {
                "required": [
                    "category_id",
                    "lang_id",
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "description": "The ID of the category",
                        "type": "integer"
                    },
                    "lang_id": {
                        "description": "The ID of the language",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the category in the specified language",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Extra": {
                "required": [
                    "hotel_id",
                    "image"
                ],
                "properties": {
                    "hotel_id": {
                        "description": "The ID of the hotel",
                        "type": "integer"
                    },
                    "image": {
                        "description": "The image of the extra",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        }
    }
}